todo-list-app/backned/requirements.txt
azure-functions
python-dateutil